requirements.txt
aiogram==3.13.1
fastapi==0.115.0
pydantic==2.9.2
sqlalchemy==2.0.35
uvicorn==0.31.0
jinja2==3.1.4
pydantic_settings==2.5.2
alembic==1.13.3
aiosqlite==0.20.0
pillow
qrcode
python-multipart